+++ db/upgrade/1643894167_update_bk_booking_settings_isvisible.php
@@ -6,7 +6,8 @@ class CoreUpgradeDB1643894167 extends Model {
   public function run(){
     Configuration::getLogger()->info("[db][upgrade] Apply update bk_booking_settings isvisible");
     $sql = "alter table bk_booking_settings modify column is_visible int NOT NULL";
-        $this->runRequest($sql);
+    $this->runRequest($sql);
+    Configuration::getLogger()->info("[db][upgrade] Apply update bk_booking_settings isvisible, done!");
   }
 }
 $db = new CoreUpgradeDB1643894167();